first few things moved around. more to move. gonna work on work comp now
